# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ВСМ-тренажёр проводника
  description: Нелинейные сценарии с таймером, шкалами и очками компетенций.
  version: 0.1.0
paths:
...
  /api/analytics/events:
    post:
      tags:
      - analytics
      summary: Записать действие пользователя на клиенте
      operationId: record_event_api_analytics_events_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientEvent'
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
                title: Response Record Event Api Analytics Events Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/analytics/me:
    get:
      tags:
      - analytics
      summary: Аналитика текущего сотрудника
      operationId: my_analytics_api_analytics_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Analytics'
      security:
      - HTTPBearer: []
  /api/analytics/team:
    get:
      tags:
      - analytics
      summary: Проводники депо инструктора (только для инструктора)
      operationId: team_api_analytics_team_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TeamMember'
                type: array
                title: Response Team Api Analytics Team Get
      security:
      - HTTPBearer: []
  /api/analytics/employees/{employee_id}:
    get:
      tags:
      - analytics
      summary: Аналитика проводника своего депо (только для инструктора)
      operationId: employee_analytics_api_analytics_employees__employee_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: employee_id
        in: path
        required: true
        schema:
          type: integer
          title: Employee Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Analytics'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Analytics:
      properties:
        employee_id:
          type: integer
          title: Employee Id
        full_name:
          type: string
          title: Full Name
        total_runs:
          type: integer
          title: Total Runs
        unfinished_runs:
          type: integer
          title: Unfinished Runs
        completion_rate:
          anyOf:
          - type: number
          - type: 'null'
          title: Completion Rate
        avg_decision_seconds:
          anyOf:
          - type: number
          - type: 'null'
          title: Avg Decision Seconds
        progress:
          items:
            $ref: '#/components/schemas/WeekPoint'
          type: array
          title: Progress
        categories:
          items:
            $ref: '#/components/schemas/CategoryStats'
          type: array
          title: Categories
        competences:
          items:
            $ref: '#/components/schemas/CompetenceStat'
          type: array
          title: Competences
        strengths:
          items:
            type: string
          type: array
          title: Strengths
        weaknesses:
          items:
            type: string
          type: array
          title: Weaknesses
        mistakes:
          items:
            type: string
          type: array
          title: Mistakes
        recommendation:
          anyOf:
          - $ref: '#/components/schemas/Recommendation'
          - type: 'null'
      type: object
      required:
      - employee_id
      - full_name
      - total_runs
      - unfinished_runs
      - completion_rate
      - avg_decision_seconds
      - progress
      - categories
      - competences
      - strengths
      - weaknesses
      - mistakes
      - recommendation
      title: Analytics
...
    CategoryStats:
      properties:
        category:
          type: string
          title: Category
        title:
          type: string
          title: Title
        runs:
          type: integer
          title: Runs
        decisions:
          type: integer
          title: Decisions
        timed_steps:
          type: integer
          title: Timed Steps
        success_rate:
          anyOf:
          - type: number
          - type: 'null'
          title: Success Rate
        best_choice_rate:
          anyOf:
          - type: number
          - type: 'null'
          title: Best Choice Rate
        timeout_rate:
          anyOf:
          - type: number
          - type: 'null'
          title: Timeout Rate
        average_reaction_share:
          anyOf:
          - type: number
          - type: 'null'
          title: Average Reaction Share
      type: object
      required:
      - category
      - title
      - runs
      - decisions
      - timed_steps
      - success_rate
      - best_choice_rate
      - timeout_rate
      - average_reaction_share
      title: CategoryStats
...
    ClientEvent:
      properties:
        type:
          type: string
          enum:
          - debrief_opened
          - notification_opened
          - run_exited
          title: Type
        run_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Run Id
        notification_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Notification Id
        platform:
          type: string
          enum:
          - web
          - android
          title: Platform
      type: object
      required:
      - type
      - platform
      title: ClientEvent
      description: 'Действие, которое видит только клиент: сервер не знает, открыт ли экран или закрыто ли приложение.'
...
    CompetenceStat:
      properties:
        code:
          type: string
          title: Code
        title:
          type: string
          title: Title
        points:
          type: integer
          title: Points
      type: object
      required:
      - code
      - title
      - points
      title: CompetenceStat
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Recommendation:
      properties:
        scenario_id:
          type: string
          title: Scenario Id
        title:
          type: string
          title: Title
        reason:
          type: string
          title: Reason
      type: object
      required:
      - scenario_id
      - title
      - reason
      title: Recommendation
...
    TeamMember:
      properties:
        employee_id:
          type: integer
          title: Employee Id
        full_name:
          type: string
          title: Full Name
        brigade:
          type: string
          title: Brigade
        level_title:
          type: string
          title: Level Title
        xp:
          type: integer
          title: Xp
        last_activity_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Activity At
        weakest_competence:
          anyOf:
          - type: string
          - type: 'null'
          title: Weakest Competence
      type: object
      required:
      - employee_id
      - full_name
      - brigade
      - level_title
      - xp
      - last_activity_at
      - weakest_competence
      title: TeamMember
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    WeekPoint:
      properties:
        week_start:
          type: string
          format: date-time
          title: Week Start
        xp:
          type: integer
          title: Xp
        runs:
          type: integer
          title: Runs
        successes:
          type: integer
          title: Successes
      type: object
      required:
      - week_start
      - xp
      - runs
      - successes
      title: WeekPoint
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer